Seperated lambdaDemo into its own component file
